Separate classes to different sub-packages.
