+++ src/main/java/com/parser/utils/OffersManagerImpl.java
@@ -1,6 +1,7 @@
 package com.parser.utils;
 
 import com.google.inject.Inject;
+import com.parser.entities.Offer;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
